Forced output data types.

+++ src/flopa/io/ptuio/reconstructor.py
@@ -239,7 +239,7 @@ class ImageReconstructor:
                 len(self.config.line_accumulations),
                 max(self.active_detectors) + 1,
                 self.tcspc_channels
-            ))
+            ), dtype=np.uint64)
             tcspc_histogram[:,0,:,:] = self.tcspc_hist[:,:detectors,:]
             data["tcspc_histogram"] = (("frame","sequence","detector","tcspc_channel"),tcspc_histogram)
 
@@ -251,7 +251,7 @@ class ImageReconstructor:
                 self.config.lines,
                 self.config.pixels,
                 max(self.active_detectors) + 1
-            ))
+            ), dtype = np.uint32)
 
             pattern = np.repeat(np.arange(len(self.config.line_accumulations)), self.config.line_accumulations)
 
@@ -264,7 +264,7 @@ class ImageReconstructor:
 
 
         if "arrival_sum" in self._required:
-            arrival_sum = np.zeros_like(photon_count)
+            arrival_sum = np.zeros_like(photon_count, dtype = np.float32)
 
         if "phasor_sum" in self._required:
             phasor_sum = np.zeros_like(photon_count, dtype=np.complex64)
@@ -289,7 +289,7 @@ class ImageReconstructor:
                     
         if "mean_arrival_time" in self.requested_outputs:
             with np.errstate(divide='ignore', invalid='ignore'):
-                mean_arrival = np.true_divide(arrival_sum, photon_count)
+                mean_arrival = np.true_divide(arrival_sum, photon_count, dtype=np.float32)
                 mean_arrival[photon_count == 0] = 0  # set empty pixels to 0
             data["mean_arrival_time"] = (("frame", "sequence", "line", "pixel", "detector"), mean_arrival)
         
@@ -298,7 +298,7 @@ class ImageReconstructor:
 
             # Normalize phasor
             with np.errstate(divide='ignore', invalid='ignore'):
-                norm_phasor = np.true_divide(phasor_sum, photon_count)
+                norm_phasor = np.true_divide(phasor_sum, photon_count, dtype=np.complex64)
                 # norm_phasor[photon_count == 0] = 0
                 norm_phasor[photon_count == 0] = np.nan + 1j * np.nan
